refactor(rengine/macros): factorize struct fields macros
